refactor: changed the file system code to use a controller and have cleaner routes
